Guest skill: replace Beginner/Intermediate/Advanced enum with float32 [0,1]
Tiers are 0–0.33 (beginner), 0.33–0.66 (intermediate), 0.66+ (advanced).
Glade preference is now restricted to advanced guests (skill >= 0.66);
intermediate and beginner guests always avoid trees.

Also adds SNOW_IMPROVEMENTS.md with design notes for the upcoming
discrete snow-type / transition-matrix redesign.

+++ internal/save/format.go
@@ -193,7 +193,7 @@ type GuestData struct {
 	ID              uint64  `json:"id,omitempty"`
 	Name            string  `json:"name,omitempty"`
 	Discipline      uint8   `json:"disc,omitempty"`
-	Skill           uint8   `json:"skill,omitempty"`
+	Skill           float32 `json:"skill,omitempty"`
 	VisitsPerSeason float32 `json:"vps,omitempty"`
 	LikesGlades     bool    `json:"glades,omitempty"`
 	PrefersGroomed  bool    `json:"groomed,omitempty"`
